# Improve performance of `Base.Fix1` and `Base.Fix2`