--- a/Project.toml
+++ b/Project.toml
@@ -43,6 +43,7 @@ Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
 LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
+StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c" # otherwise we can't add a compat bound
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]