This PR addresses https://github.com/FluxML/Zygote.jl/issues/957.

Currently (copied from the issue above):
```julia
julia> using Zygote, BenchmarkTools

julia> x = rand(100);

julia> @btime Zygote.gradient($(x -> prod(y -> y + 1, x)), $x);
  10.373 μs (343 allocations: 13.27 KiB)

julia> @btime Zygote.gradient($(x -> prod(Base.Fix1(+, 1), x)), $x);
  342.138 μs (3608 allocations: 138.08 KiB)

julia> @btime Zygote.gradient($(x -> prod(Base.Fix2(+, 1), x)), $x);
  352.705 μs (3607 allocations: 138.06 KiB)
```

With this PR:
```julia
julia> using Zygote, BenchmarkTools

julia> x = rand(100);

julia> @btime Zygote.gradient($(x -> prod(y -> y + 1, x)), $x);
  10.033 μs (343 allocations: 13.27 KiB)

julia> @btime Zygote.gradient($(x -> prod(Base.Fix1(+, 1), x)), $x);
  12.150 μs (453 allocations: 27.73 KiB)

julia> @btime Zygote.gradient($(x -> prod(Base.Fix2(+, 1), x)), $x);
  12.223 μs (453 allocations: 27.73 KiB)
```

My main questions are:
- Why is it still slower and causes many more allocations?
- Why are `Base.Fix1` and `Base.Fix2` so slow in the first place and don't just work?