@CarloLucibello 2021-04-30
I don't know the answer to any of those questions, but I like this simple fix to a huge performance hit.