@mcabbott 2021-05-06
For me the `@elapsed` times are backwards, and hence the timing test will pass right now. 
```
julia> x = rand(100);

julia> @btime Zygote.gradient($(x -> prod(y -> y + 1, x)), $x);
  8.833 μs (343 allocations: 13.27 KiB)

julia> @btime Zygote.gradient($(x -> prod(Base.Fix1(+, 1), x)), $x);
  266.625 μs (3608 allocations: 138.08 KiB)

julia> t = @elapsed(gradient(x -> prod(y -> y + 1, x), x))  # not the first run
0.214923125

julia> tfix = @elapsed(gradient(x -> prod(Base.Fix1(+, 1), x), x))
0.029850208

julia> tfix < 2t  # is the test
true

(@v1.6) pkg> st Zygote  # similar on master
  [e88e6eb3] Zygote v0.6.10

julia> versioninfo()
Julia Version 1.6.0
Commit f9720dc2eb (2021-03-24 12:55 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin19.6.0) 
  CPU: Apple M1
```
Likewise allocations:
```
julia> @ballocated(gradient(x -> prod(y -> y + 1, x), $x))
13584

julia> @ballocated(gradient(x -> prod(Base.Fix1(+, 1), x), $x))
141392

julia> @allocated(gradient(x -> prod(y -> y + 1, x), x))
53513519

julia> @allocated(gradient(x -> prod(Base.Fix1(+, 1), x), x))
4279117```