@devmotion 2021-05-06
With this PR I obtain:

```julia
julia> using Zygote, BenchmarkTools

julia> const x = rand(100);

julia> @btime Zygote.gradient($(x -> prod(y -> y + 1, x)), $x);
  10.151 μs (343 allocations: 13.27 KiB)

julia> @btime Zygote.gradient($(x -> prod(Base.Fix2(+, 1), x)), $x);
  12.284 μs (453 allocations: 27.73 KiB)

julia> @elapsed(gradient(x -> prod(y -> y + 1, x), x))
0.338051284

julia> @elapsed(gradient(x -> prod(Base.Fix2(+, 1), x), x))
0.064180138

julia> @ballocated(gradient($(x -> prod(y -> y + 1, x)), $x))
13584

julia> @ballocated(gradient($(x -> prod(Base.Fix2(+, 1), x)), $x))
28400

julia> @allocated(gradient(x -> prod(y -> y + 1, x), x))
53453284

julia> @allocated(gradient(x -> prod(Base.Fix2(+, 1), x), x))
3991125

(Zygote) pkg> st
     Project Zygote v0.6.11
      Status `~/.julia/dev/Zygote/Project.toml`
  [621f4979] AbstractFFTs v1.0.1
  [6e4b80f9] BenchmarkTools v0.7.0
  [082447d4] ChainRules v0.7.61
  [d360d2e6] ChainRulesCore v0.9.41
  [b552c78f] DiffRules v1.0.2
  [1a297f60] FillArrays v0.11.7
  [f6369f11] ForwardDiff v0.10.18
  [7869d1d1] IRTools v0.4.2
  [1914dd2f] MacroTools v0.5.6
  [77ba4419] NaNMath v0.3.5
  [ae029012] Requires v1.1.3
  [276daf66] SpecialFunctions v1.3.0
  [700de1a5] ZygoteRules v0.2.1
  [8ba89e20] Distributed
  [b77e0a4c] InteractiveUtils
  [37e2e46d] LinearAlgebra
  [9a3f8284] Random
  [10745b16] Statistics

julia> versioninfo()
Julia Version 1.6.1
Commit 6aaedecc44 (2021-04-23 05:59 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: Intel(R) Core(TM) i7-7500U CPU @ 2.70GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-11.0.1 (ORCJIT, skylake)
Environment:
  JULIA_NUM_THREADS = 4
```

The differences seem to be caused between `@elapsed` and `@btime` and `@allocated` and `@ballocated` seem to be caused by the fact more time and allocations have to be spent for constructing the closure:
```julia
julia> f(x) = prod(y -> y + 1, x)
f (generic function with 1 method)

julia> g(x) = prod(Base.Fix2(+, 1), x)
g (generic function with 1 method)

julia> @elapsed(gradient(f, x))
0.0001255

julia> @elapsed(gradient(g, x))
0.000140757

julia> @allocated(gradient(f, x))
13568

julia> @allocated(gradient(g, x))
28384
```

In any case, probably better to use `@ballocated` and `@btime` here.