@mcabbott 2021-05-06
> In any case, probably better to use `@ballocated` and `@btime` here.

With the downside that these make tests take longer, and they already take quite a while. My vote would probably be just to omit the test.