@devmotion 2021-05-06
I don't know, the main point here is to fix the time it takes to compute the gradient. There's no problem with the value of the gradient, it's computed just fine by Zygote automatically, but it is very slow. So the only way to check for the problem that is fixed by the PR seems to be to add tests with `@belapsed`.