@devmotion 2021-05-06
> With the downside that these make tests take longer, and they already take quite a while.

If the tests take too long, maybe it would be helpful to split them into different groups (similar to e.g. https://github.com/SciML/OrdinaryDiffEq.jl/blob/master/test/runtests.jl).