@mcabbott 2021-05-06
Right. But there are lots of potential performance problems, and it's hard to test them all. We don't have a comprehensive set of benchmarks to test for regressions, but even if we did, they would probably be too expensive to run on every commit. Julia itself has such a set, but they aren't included in the normal tests, I think for similar reasons -- they take ages, can't be done accurately on random CI hardware.