@devmotion 2021-05-06
Sure, one can't test everything, but this is a PR that is really only about major performance issues. These issues were discovered in KernelFunctions only accidentally but affect many other packages and users. The rewriting of Fix1 and Fix2 in this PR should not be needed at all and in principle it is completely stupid to add these rules - but it fixes the performance problem to some extent. So I think it should definitely be tested to ensure that these problems don't reappear. Such tests with BenchmarkTools work reliably e.g. in RuntimeGeneratedFunctions.