@mcabbott 2021-05-06
Yes, I hear you. If someone else thinks it's OK to start using BenchmarkTools, they can merge it.