@DhairyaLGandhi 2021-06-14
Adding BenchmarkTools isn't going to help our already length CI times, so I would omit for now. We can try to add a benchmark repo separately (FluxBench already exists) where we unclog CI here and track performance too.