@devmotion 2021-06-14
I disabled the regression tests since I think it would be good to release this fix and I don't want to block the PR.