@ChrisRackauckas 2021-06-16
>Right. But there are lots of potential performance problems, and it's hard to test them all. We don't have a comprehensive set of benchmarks to test for regressions, but even if we did, they would probably be too expensive to run on every commit. Julia itself has such a set, but they aren't included in the normal tests, I think for similar reasons -- they take ages, can't be done accurately on random CI hardware.

Base is doing something that isn't relevant here. It's testing things like the speed of `sin`, which need to be extremely accurate. When you're trying to figure out regressions in something that takes 10ns, you need a ton of samples. Even though it's faster, it's slower because it's harder. Even having an email client open on Nanosoldier will make those tests fail because they are so close to the metal. As things get larger, it's simpler and simpler to just take one or two calls and get a representative sample.